% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annohub.R
\name{annohub}
\alias{annohub}
\title{Get AnnotationHub object}
\usage{
annohub(
  ...,
  version = NULL,
  hub = NULL,
  cache = NULL,
  proxy = NULL,
  local = NULL
)
}
\arguments{
\item{...}{See \code{\link[AnnotationHub:AnnotationHub-class]{AnnotationHub}} for additional
arguments.}

\item{version}{A valid Bioconductor version, see:
\url{https://bioconductor.org/config.yaml} for all version.}

\item{hub}{The URL for the online AnnotationHub.}

\item{cache}{The file system location of the local AnnotationHub cache.}

\item{proxy}{Set the proxy.}

\item{local}{A bool, whether to use the cache.}
}
\value{
A \code{\link[AnnotationHub:AnnotationHub-class]{AnnotationHub}} instance.
}
\description{
\code{\link[AnnotationHub:AnnotationHub-class]{AnnotationHub()}} restrict the
AnnotationHub snapshotDate before the Bioconductor version, this function
just bypass the restriction to let you query all AnnotationHub data.
}
