% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enricher-package.R
\docType{package}
\name{enricher-package}
\alias{enricher}
\alias{enricher-package}
\title{enricher: Lightweight Functional Enrichment for Omics Workflows}
\description{
Provides a lightweight and extensible interface for performing functional enrichment analysis of gene sets to support the interpretation of omics data. Supports multiple annotation resources, including KEGG and other biological databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/enricher}
  \item Report bugs at \url{https://github.com/Yunuuuu/enricher/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
