% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset-methods.R
\name{geneset}
\alias{geneset}
\alias{geneset.default}
\alias{geneset.character}
\title{Create a single geneset}
\usage{
geneset(geneset, ...)

\method{geneset}{default}(geneset, id = NULL, term = NULL, description = NULL, ...)

\method{geneset}{character}(geneset, id = NULL, term = NULL, description = NULL, ...)
}
\arguments{
\item{geneset}{A character of geneset elements.}

\item{...}{Additional attributes to be added to the geneset.}

\item{id}{A single string of the geneset ID.}

\item{term}{A single string of the geneset term.}

\item{description}{A single string of the geneset description.}
}
\description{
Create a single geneset
}
