% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-method.R
\name{genesets}
\alias{genesets}
\alias{genesets.data.frame}
\alias{genesets.list}
\alias{genesets.character}
\alias{genesets.OrgDb}
\title{Create Gene Sets}
\usage{
genesets(gs, ...)

\method{genesets}{data.frame}(gs, ..., select = NULL)

\method{genesets}{list}(gs, ..., ids = NULL, terms = NULL, descriptions = NULL)

\method{genesets}{character}(gs, ..., verbose = TRUE)

\method{genesets}{OrgDb}(gs, ..., keytype = "SYMBOL", ontology = NULL)
}
\arguments{
\item{gs}{Specify the gene sets to use. Accepted formats:
\itemize{
\item \code{character}: A string indicating the source of gene sets. Acceptable values
include \code{"go/GO"} or \code{"kegg/KEGG"}, which will download gene
sets from the \code{GO} or \code{KEGG} databases. You may also specify the name of an
\href{https://bioconductor.org/packages/release/BiocViews.html#___OrgDb}{OrgDb}
package, which will call the \code{OrgDb} method. Alternatively, provide a file
path to a \code{.gmt} or \code{.gmx} file to import custom gene sets.
\item \code{data.frame}: A data frame with \code{2}, \code{3}, or \code{4} columns, representing gene
set IDs, optional terms (if \code{4} columns), optional descriptions (if
\code{3}/\code{4} columns), and associated genes or features.
\item \code{list}: A list where each element is a character vector representing a gene
set. A \code{term}/\code{description} attribute can be attached to each element, or a
\code{terms}/\code{descriptions} attribute to the entire list.
\item \code{OrgDb}: An
\href{https://bioconductor.org/packages/release/BiocViews.html#___OrgDb}{OrgDb}
object used to retrieve GO gene sets.
}}

\item{...}{Additional arguments passed to the method.}

\item{select}{A character or integer vector select the columns to be used for
a data.frame input.}

\item{ids}{A character vector of gene set IDs for list input.}

\item{terms}{A character vector of gene set terms for list input.}

\item{descriptions}{A character vector of gene set descriptions for a list
input.}

\item{verbose}{A single logical value indicates whether the process should be
verbose.}

\item{keytype}{A single character string indicating which type of keys to
use.}

\item{ontology}{A character vector specifying the GO ontology to use. Must be
one or more of \code{"BP"}, \code{"CC"}, and \code{"MF"}.}
}
\value{
A \code{enricher_genesets} object.
}
\description{
Create Gene Sets
}
