% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gsea_gene_permutate}
\alias{gsea_gene_permutate}
\title{GSEA gene permutation analysis.}
\usage{
gsea_gene_permutate(identifiers, metrics, genesets, exponent, nperm, seed)
}
\description{
GSEA gene permutation analysis.
}
\section{Arguments}{
\itemize{
\item \code{identifiers}: Character vector of gene identifiers.
\item \code{metrics}: Numeric vector of ranking metrics.
\item \code{genesets}: List of character vectors, each representing a geneset.
\item \code{exponent}: Weighting exponent for metrics.
\item \code{nperm}: Number of permutations.
}
}

\section{Returns}{
A list containing enrichment results.
}

