% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-kegg.R
\name{kegg_pathway}
\alias{kegg_pathway}
\title{Retrieve KEGG Gene Sets or Pathways}
\usage{
kegg_pathway(link = "hsa", strategy = NULL, save = NULL, verbose = TRUE)
}
\arguments{
\item{link}{A KEGG organism code (e.g., \code{"hsa"} for human). A full list is
available via \code{keggdb("organism")$organism}. This will link pathways using
gene symbols for the specified organism. Alternatively, a KEGG database name
can be provided to link pathway associations based on database entry IDs.

\strong{Note}: the following databases can be linked to pathways:
\code{"module"}, \code{"ko"}, \code{"compound"}, \code{"glycan"}, \code{"reaction"}, \code{"rclass"}, \code{"enzyme"}, \code{"disease"}, and \code{"drug"}.}

\item{strategy}{Character string specifying how to access the database and
whether to use caching:
\itemize{
\item \code{"read"}: Read from the saved file only. Error if not exist.
\item \code{"download"}: Download fresh data only. Do not save it.
\item \code{"save"}: Download data and save it locally.
}

By default, \code{NULL} is used, which means read from saved file if available;
otherwise, download and save the data.}

\item{save}{A string specifying the file path to read from or save to. File
extension \code{".rds"} will be automatically added.}

\item{verbose}{A single logical value indicates whether the process should be
verbose.}
}
\value{
A data frame
}
\description{
Retrieves gene sets or pathways from the KEGG database for a given organism.
}
