% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-kegg.R
\name{keggdb}
\alias{keggdb}
\title{Get database from KEGG}
\usage{
keggdb(database, organism = NULL, strategy = NULL, save = NULL, verbose = TRUE)
}
\arguments{
\item{database}{A KEGG database (list available via
\code{\link[KEGGREST:listDatabases]{listDatabases()}}). In addition, \code{"organism"} can
be used to retrieve all available organisms in the KEGG database.}

\item{organism}{A KEGG organism code (list available via
\code{keggdb("organism")$organism}.}

\item{strategy}{Character string specifying how to access the database and
whether to use caching:
\itemize{
\item \code{"read"}: Read from the saved file only. Error if not exist.
\item \code{"download"}: Download fresh data only. Do not save it.
\item \code{"save"}: Download data and save it locally.
}

By default, \code{NULL} is used, which means read from saved file if available;
otherwise, download and save the data.}

\item{save}{A string specifying the file path to read from or save to. File
extension \code{".rds"} will be automatically added.}

\item{verbose}{A single logical value indicates whether the process should be
verbose.}
}
\value{
A list of contents for each entries.
}
\description{
Get database from KEGG
}
