% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-kegg.R
\name{keggdb_find}
\alias{keggdb_find}
\title{Finds entries in a given database}
\usage{
keggdb_find(database, ..., option = NULL)
}
\arguments{
\item{database}{
  Either the name of a single KEGG database (list available via
  \code{\link[KEGGREST]{listDatabases}()}, a "T number" genome identifier,
  or a KEGG organism code (lists of both available via
  \code{keggList("organism")}).
}

\item{...}{One or more keywords, or a range of integers representing
molecular weights. If identifiers included are not known to KEGG, the
results will not contain any information about those identifiers.}

\item{option}{
    \code{Optional.} If \code{database} is \code{compound} or \code{drug},
    \code{option} can be \code{formula}, \code{exact_mass}, or 
    \code{weight}. 
    Chemical formula search is a partial match irrespective of the
    order of atoms given. 
    The exact mass (or molecular weight) is checked by rounding off to the
    same decimal place as the query data.
}
}
\description{
Finds entries in a given database
}
