% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-kegg.R
\name{keggdb_get}
\alias{keggdb_get}
\title{Retrieves given database entries}
\usage{
keggdb_get(..., option = NULL)
}
\arguments{
\item{...}{One or more KEGG identifiers.}

\item{option}{
    \code{Optional.} Option governing the format of the output.
    \code{aaseq} is an amino acid sequence, \code{ntseq} is a nucleotide
    sequence. \code{image} returns an object which can be written
    to a PNG file, \code{kgml} returns a KGML document.
}
}
\description{
Retrieves given database entries
}
