% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr-matrix.R
\name{repr_matrix}
\alias{repr_matrix}
\title{Extract matrix}
\usage{
repr_matrix(matrix, ..., `_arg` = NULL)
}
\value{
A \code{xbio_matrix} object, which is a named numeric vector.
}
\description{
Extract matrix
}
