% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr-metrics.R
\name{repr_metrics}
\alias{repr_metrics}
\alias{repr_metrics.numeric}
\alias{repr_metrics.data.frame}
\alias{repr_metrics.character}
\title{Create ranking metrics representation}
\usage{
repr_metrics(metrics, ..., `_arg` = NULL)

\method{repr_metrics}{numeric}(metrics, ..., `_arg` = NULL)

\method{repr_metrics}{data.frame}(metrics, ..., `_arg` = NULL)

\method{repr_metrics}{character}(metrics, ..., `_arg` = NULL)
}
\arguments{
\item{metrics}{A named numeric ranked metrics.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A named numeric vector with class \code{xbio_metrics}.
}
\description{
Create ranking metrics representation
}
