% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbio-package.R
\docType{package}
\name{xbio-package}
\alias{xbio}
\alias{xbio-package}
\title{xbio: Cross-Process Scoring and Linking Framework for Interpreting Biological Associations}
\description{
An R package designed to bridge functional relationships between biological processes through flexible and extensible scoring strategies. It enables users to quantify associations between two sets of processes—whether well-characterized, poorly understood, or both—based on gene sets, ranked metrics, or threshold-defined signatures. By supporting multiple enrichment and similarity-based methods, including over-representation analysis (ORA), GSEA variants, and rank-rank hypergeometric overlap (RRHO), 'xbio' provides a unified interface for inferring potential functional relevance, co-regulation, or shared biological roles. This makes it a powerful tool for functional interpretation, knowledge transfer, and hypothesis generation in genomics, transcriptomics, and systems biology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/xbio}
  \item Report bugs at \url{https://github.com/Yunuuuu/xbio/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
