% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge-.R
\name{bridge}
\alias{bridge}
\title{Bridge Functional Links Between Biological Processes}
\usage{
bridge(source, target, method)
}
\arguments{
\item{source}{The input source to be scored, depending on the selected
\code{method}.
\itemize{
\item \strong{Over-Representation Analysis}: \code{\link[=repr_threshold]{repr_threshold()}}.
\item \strong{GSEA gene-permutation algorithm}: \code{\link[=repr_metrics]{repr_metrics()}}.
\item \strong{GSEA sample-permutation algorithm}: Accepts a matrix-like object.
Although this method inherently relies on a ranked metric, it internally
computes the ranking metric and performs permutations based on sample
labels. Metric-related parameters should be specified within the \code{method}
argument.

\code{Note}: While matrices are not conceptually equivalent to biological
process representations, we provide \code{\link[=repr_matrix]{repr_matrix()}} for stylistic
consistency and user convenience.
\item \strong{Rank-Rank Hypergeometric Overlap test}: \code{\link[=repr_metrics]{repr_metrics()}}.
}}

\item{target}{The input target used for associating, depending on the
selected \code{method}.
\itemize{
\item \strong{Over-Representation Analysis}: \code{\link[=repr_genesets]{repr_genesets()}}.
\item \strong{GSEA gene-permutation algorithm}: \code{\link[=repr_genesets]{repr_genesets()}}.
\item \strong{GSEA sample-permutation algorithm}: \code{\link[=repr_genesets]{repr_genesets()}}.
\item \strong{Rank-Rank Hypergeometric Overlap test}: \code{\link[=repr_metrics]{repr_metrics()}}.
}}

\item{method}{The scoring method to use:
\itemize{
\item \strong{Over-Representation Analysis}: \code{\link[=ORA]{ORA()}}
\item \strong{GSEA gene-permutation algorithm}: \code{\link[=GSEAGene]{GSEAGene()}}, \code{\link[=GSEASimple]{GSEASimple()}},
\code{\link[=GSEAMultilevel]{GSEAMultilevel()}}, and \code{\link[=GSEABroadGene]{GSEABroadGene()}}.
\item \strong{GSEA sample-permutation algorithm}: \code{\link[=GSEABroad]{GSEABroad()}} and \code{\link[=GSEASample]{GSEASample()}}.
\item \strong{Rank-Rank Hypergeometric Overlap test}: \code{\link[=RRHO]{RRHO()}}.
}}
}
\description{
In biological research, scoring methods are commonly used to evaluate the
association between two sets of biological processes. These may include
well-characterized processes (the \emph{target}), poorly understood ones (the
\emph{source}), or even two processes of unknown function. By quantifying their
association, we can infer potential biological relevance or shared
functionality. A strong association suggests functional similarity or
coordinated regulation.

Biological processes are typically represented by one of the following:
\itemize{
\item One or more gene sets (\code{\link[=repr_genesets]{repr_genesets()}})
\item A ranked metric (\code{\link[=repr_metrics]{repr_metrics()}})
\item A gene set based on a hard threshold (\code{\link[=repr_threshold]{repr_threshold()}})
}
}
