% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset-class.R
\name{geneset}
\alias{geneset}
\alias{geneset.character}
\alias{geneset.default}
\title{Create a single geneset}
\usage{
geneset(id, geneset, term = NULL, description = NULL, ...)

\method{geneset}{character}(id, geneset, term = NULL, description = NULL, ...)

\method{geneset}{default}(id, geneset, term = NULL, description = NULL, ...)
}
\arguments{
\item{id}{A single string of the geneset ID.}

\item{geneset}{A character of geneset elements.}

\item{term}{A single string of the geneset term.}

\item{description}{A single string of the geneset description.}

\item{...}{Additional attributes to be added to the geneset.}
}
\description{
Create a single geneset
}
