% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-gs.R
\name{gs_map}
\alias{gs_map}
\alias{gs_biomart}
\title{Map a geneset or genesets object to a different keytype}
\usage{
gs_map(gs, annodb, key_source, key_target, ...)

gs_biomart(gs, mart, key_source, key_target, ...)
}
\arguments{
\item{gs}{A \code{\link[=geneset]{geneset()}} or \code{\link[=genesets]{genesets()}} object.}

\item{annodb}{An \code{AnnotationDb} object or a character string naming the
annotation package.}

\item{key_source}{A string indicating the keytype used in \code{gs}.}

\item{key_target}{A string indicating the desired target keytype.}

\item{...}{Additional arguments passed to
\code{\link[AnnotationDbi:AnnotationDb-class]{mapIds()}} or \code{\link[biomaRt:getBM]{getBM()}}.}

\item{mart}{A \code{\link[biomaRt:useMart]{Mart}} object.}
}
\value{
An object of the same class as \code{gs}, with mapped gene identifiers.
}
\description{
This function maps the gene identifiers in a \code{\link[=geneset]{geneset()}} or \code{\link[=genesets]{genesets()}}
object from one keytype to another, using either an
\code{\link[AnnotationDbi:AnnotationDb-class]{AnnotationDb}} or a \code{\link[biomaRt:useMart]{Mart}}
object.
}
