% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge-gsea-prerank.R
\name{GSEAPrerank}
\alias{GSEAPrerank}
\alias{GSEAGene}
\alias{GSEASimple}
\alias{GSEAMultilevel}
\title{GSEA prerank algorithm}
\usage{
GSEAPrerank(exponent = 1, nperm = 1000L)

GSEAGene(exponent = 1, nperm = 1000, threads = NULL)

GSEASimple(exponent = 1, nperm = 1000L, score_type = "std", threads = 1L)

GSEAMultilevel(
  eps = 1e-50,
  exponent = 1,
  nperm = 1000L,
  sample_size = 101L,
  score_type = "std",
  threads = 1L
)
}
\description{
GSEA prerank algorithm
}
\section{Functions}{
\itemize{
\item \code{GSEAPrerank()}: R-based preranked GSEA algorithm (official). See
\code{\link[GSEA:GSEA]{GSEA}} for details.

\item \code{GSEAGene()}: Rust-based preranked GSEA algorithm.

\item \code{GSEASimple()}: \code{fgsea} algorithm. See
\code{\link[fgsea:fgseaSimple]{fgseaSimple}} for details.

\item \code{GSEAMultilevel()}: \code{fgsea} algorithm. See
\code{\link[fgsea:fgseaMultilevel]{fgseaMultilevel}} for details.

}}
