% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-kegg.R
\name{keggdb}
\alias{keggdb}
\title{Get database from KEGG}
\usage{
keggdb(database, organism = NULL, strategy = NULL, verbose = TRUE)
}
\arguments{
\item{database}{A KEGG database (list available via
\code{\link[KEGGREST:listDatabases]{listDatabases()}}). In addition, \code{"organism"} can
be used to retrieve all available organisms in the KEGG database, and
\code{"genesets"} can be used to obtain gene sets for functional enrichment
analysis.}

\item{organism}{A KEGG database (list available via
\code{keggdb("organism")$organism}.}

\item{strategy}{Character string specifying how to access the database and
whether to use caching:
\itemize{
\item \code{"read"}: Read from cache only. Error if not cached.
\item \code{"download"}: Download fresh data. Do not cache it.
\item \code{"cache"}: Download data and cache it locally.
}

By default, \code{NULL} is used, which means read from cache if available;
otherwise, download and cache the data.}

\item{verbose}{A single logical value indicates whether the process should be
verbose.}
}
\value{
A list of contents for each entries.
}
\description{
Get database from KEGG
}
