% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr-genesets.R
\name{repr_genesets}
\alias{repr_genesets}
\alias{genesets}
\alias{genesets,data.frame-method}
\alias{genesets,list-method}
\alias{genesets,character-method}
\alias{genesets,OrgDb-method}
\title{Create Gene Sets representation}
\usage{
repr_genesets(gs, ..., min_size = 5, max_size = 500)

genesets(gs, ..., `_arg` = NULL)

\S4method{genesets}{data.frame}(gs, ..., select = NULL, `_arg` = NULL)

\S4method{genesets}{list}(gs, ..., ids = NULL, terms = NULL, descriptions = NULL, `_arg` = NULL)

\S4method{genesets}{character}(gs, ..., verbose = TRUE, `_arg` = NULL)

\S4method{genesets}{OrgDb}(gs, ..., keytype = "SYMBOL", ontology = NULL, `_arg` = NULL)
}
\arguments{
\item{gs}{Specify the gene sets to use. Accepted formats:
\itemize{
\item \code{character}: A string indicating the source of gene sets.
\itemize{
\item Acceptable values include \code{"go/GO"} or \code{"kegg/KEGG"}, which
download gene sets from the GO or KEGG databases. In this case,
additional arguments such as \code{organism}, \code{strategy}, and \code{cache} can be
used. See \code{\link[=keggdb]{keggdb()}} and \code{\link[=godb]{godb()}} for details.
\item You may also specify the name of an
\href{https://bioconductor.org/packages/release/BiocViews.html#___OrgDb}{\code{OrgDb}}
package, which will call the \code{OrgDb} method.
\item Alternatively, provide a file path to a \code{.gmt} or \code{.gmx} file to import
custom gene sets.
}
\item \code{data.frame}: A data frame with \code{2}, \code{3}, or \code{4} columns, representing gene
set IDs, optional terms (if \code{4} columns), optional descriptions (if
\code{3}/\code{4} columns), and associated genes or features.
\item \code{list}: A list where each element is a character vector representing a gene
set. A \code{term}/\code{description} attribute can be attached to each element, or a
\code{terms}/\code{descriptions} attribute to the entire list.
\item \code{OrgDb}: An
\href{https://bioconductor.org/packages/release/BiocViews.html#___OrgDb}{\code{OrgDb}}
object used to retrieve GO gene sets.
}}

\item{...}{Additional arguments passed to the method.}

\item{min_size, max_size}{An integer specifying the minimum/maximum size of
gene sets to include in the analysis.}

\item{_arg}{A single string of the argument name for \code{gs}, used by the
internal to provide better message.}

\item{select}{A character or integer vector select the columns to be used for
a data.frame input.}

\item{ids}{A character vector of gene set IDs for list input. If \code{NULL}, the
names of the input list will be used.}

\item{terms}{A character vector of gene set terms for list input.}

\item{descriptions}{A character vector of gene set descriptions for a list
input.}

\item{verbose}{A single logical value indicates whether the process should be
verbose.}

\item{keytype}{A single character string indicating which type of keys to
use.}

\item{ontology}{A character vector specifying the GO ontology to use. Must be
one or more of \code{"BP"}, \code{"CC"}, and \code{"MF"}.}
}
\value{
A \code{enricher_genesets} object.
}
\description{
Create Gene Sets representation
}
