% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr-matrix.R
\name{repr_matrix}
\alias{repr_matrix}
\alias{repr_matrix.matrix}
\alias{repr_matrix.dgCMatrix}
\alias{repr_matrix.Seurat}
\alias{repr_matrix.SummarizedExperiment}
\alias{repr_matrix.SingleCellExperiment}
\title{Extract matrix}
\usage{
repr_matrix(matrix, ..., `_arg` = NULL)

\method{repr_matrix}{matrix}(matrix, ..., `_arg` = NULL)

\method{repr_matrix}{dgCMatrix}(matrix, ..., `_arg` = NULL)

\method{repr_matrix}{Seurat}(
  matrix,
  assay = NULL,
  ...,
  dimred = NULL,
  n_dimred = NULL,
  layer = "data",
  `_arg` = NULL
)

\method{repr_matrix}{SummarizedExperiment}(matrix, assay = NULL, ..., `_arg` = NULL)

\method{repr_matrix}{SingleCellExperiment}(
  matrix,
  assay = NULL,
  ...,
  dimred = NULL,
  n_dimred = NULL,
  `_arg` = NULL
)
}
\arguments{
\item{matrix}{Input object used to extract the matrix.}

\item{...}{Additional arguments passed to the method.}

\item{assay}{Integer scalar or string indicating which assay to use.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{layer}{Name of layer to get for \code{Seurat} object.}
}
\value{
A matrix object.
}
\description{
Extract matrix
}
