% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr-metrics.R
\name{repr_metrics}
\alias{repr_metrics}
\alias{repr_metrics,numeric-method}
\alias{repr_metrics,data.frame-method}
\alias{repr_metrics,character-method}
\title{Create ranking metrics representation}
\usage{
repr_metrics(metrics, ..., `_arg` = NULL)

\S4method{repr_metrics}{numeric}(metrics, ..., `_arg` = NULL)

\S4method{repr_metrics}{data.frame}(metrics, ..., `_arg` = NULL)

\S4method{repr_metrics}{character}(metrics, ..., `_arg` = NULL)
}
\value{
A \code{xbio_metrics} object.
}
\description{
Create ranking metrics representation
}
